pyproject.toml
[tool.pyright]
venvPath = "venv"
venv = "."
verboseOutput = false
reportConstantRedefinition = "warning"
reportUnnecessaryComparison = "warning"
reportAssertAlwaysTrue = "warning"
reportUnusedExpression = false
include = ["example_streamlit_sql"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "build/*",
    "dist/*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["example_streamlit_sql"]

[tool.vulture]
exclude = []
ignore_decorators = []
ignore_names = []
make_whitelist = false
min_confidence = 0
paths = ["example_streamlit_sql"]
sort_by_size = true
verbose = false

[tool.isort]
profile = "black"

[metadata]
description = "Example streamlit_sql"
keyword = []
classifiers = [""]
author = "Anonymous"
author_email = ""

[project]
name = "example_streamlit_sql"
version = "0.1.0"
# readme= "docs/README.md"
requires-python = ">=3.8"
license = {text = "MIT License"}
dependencies = [
    "python-dateutil",
    "pandas",
    "python-dotenv",

   "psycopg2-binary",
   "sqlalchemy",
   "sqlalchemy_utils",


   "streamlit",
   "streamlit-date-picker",
   "streamlit_searchbox",
   "streamlit-datalist",

]

[project.urls]
# homepage = "https://github.com/edkedk99/hledger-args"
# documentation = "https://edkedk99.github.io/hledger-args/"
# repository = "https://github.com/edkedk99/hledger-args"
# changelog = ""

[project.scripts]
phlc = "example_streamlit_sql.__main__:main"
